+++ src/PalDDD.Analyzers/StrategicDddAnalyzer.cs
@@ -1,6 +1,4 @@
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Immutable;
 
 namespace PalDDD.Analyzers;